refactor: Split out checks
